# backend/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: AI Tutor API
  version: 1.0.0
servers:
  - url: /api/v1
paths:
...
  /questions/{id}:
    get:
      summary: Get question
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
...
  /quiz-attempts/{attemptId}:
    get:
      summary: Get quiz attempt
      parameters:
        - in: path
          name: attemptId
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
  /quiz-attempts/{attemptId}/items/{itemId}:
    put:
      summary: Save/grade answer
      parameters:
        - in: header
          name: Idempotency-Key
          required: false
          schema: { type: string }
      responses:
        '200': { description: OK }
...